refactor provider unit tests (#250, #251)
