+++ tests/providers/fixtures.py
@@ -1,0 +1,69 @@
+# Copyright (C) 2023 qBraid
+#
+# This file is part of the qBraid-SDK
+#
+# The qBraid-SDK is free software released under the GNU General Public License v3
+# or later. You can redistribute and/or modify it under the terms of the GPL v3.
+# See the LICENSE file in the project root or <https://www.gnu.org/licenses/gpl-3.0.html>.
+#
+# THERE IS NO WARRANTY for the qBraid-SDK, as per Section 15 of the GPL v3.
+
+"""
+Helper functions for providers testing
+
+"""
+import braket
+import cirq
+import numpy as np
+import qiskit
+
+from qbraid.api import QbraidSession
+
+
+def device_wrapper_inputs(vendor: str):
+    """Returns list of tuples containing all device_wrapper inputs for given vendor."""
+    session = QbraidSession()
+    devices = session.get("/public/lab/get-devices", params={}).json()
+    input_list = []
+    deprecated = ["aws_ionq"]
+    for document in devices:
+        if document["vendor"] == vendor:
+            qbraid_id = document["qbraid_id"]
+            if qbraid_id not in deprecated:
+                input_list.append(qbraid_id)
+    return input_list
+
+
+def braket_circuit():
+    """Returns low-depth, one-qubit Braket circuit to be used for testing."""
+    circuit = braket.circuits.Circuit()
+    circuit.h(0)
+    circuit.ry(0, np.pi / 2)
+    return circuit
+
+
+def cirq_circuit(meas=True):
+    """Returns Low-depth, one-qubit Cirq circuit to be used for testing.
+    If ``meas=True``, applies measurement operation to end of circuit."""
+    q0 = cirq.GridQubit(0, 0)
+
+    def basic_circuit():
+        yield cirq.H(q0)
+        yield cirq.Ry(rads=np.pi / 2)(q0)
+        if meas:
+            yield cirq.measure(q0, key="q0")
+
+    circuit = cirq.Circuit()
+    circuit.append(basic_circuit())
+    return circuit
+
+
+def qiskit_circuit(meas=True):
+    """Returns Low-depth, one-qubit Qiskit circuit to be used for testing.
+    If ``meas=True``, applies measurement operation to end of circuit."""
+    circuit = qiskit.QuantumCircuit(1, 1) if meas else qiskit.QuantumCircuit(1)
+    circuit.h(0)
+    circuit.ry(np.pi / 2, 0)
+    if meas:
+        circuit.measure(0, 0)
+    return circuit
