+++ tests/providers/test_braket_provider.py
@@ -9,7 +9,7 @@
 # THERE IS NO WARRANTY for the qBraid-SDK, as per Section 15 of the GPL v3.
 
 """
-Unit tests for managing Quantum Jobs
+Unit tests for BraketProvider class
 
 """
 import os
@@ -17,10 +17,10 @@ import os
 import pytest
 from braket.circuits import Circuit
 
-from qbraid import device_wrapper
+from qbraid import device_wrapper, job_wrapper
 from qbraid.providers.aws import BraketProvider
 
-# Skip tests if IBM/AWS account auth/creds not configured
+# Skip tests if AWS account auth/creds not configured
 skip_remote_tests: bool = os.getenv("QBRAID_RUN_REMOTE_TESTS") is None
 REASON = "QBRAID_RUN_REMOTE_TESTS not set (requires configuration of AWS storage)"
 
@@ -54,3 +54,25 @@ def test_braket_queue_visibility():
         queue_position = job.queue_position()
         job.cancel()
         assert isinstance(queue_position, int)
+
+
+@pytest.mark.skipif(skip_remote_tests, reason=REASON)
+@pytest.mark.parametrize(
+    "company,region", [("rigetti", "us-west-1"), ("ionq", "us-east-1"), ("oqc", "eu-west-2")]
+)
+def test_get_region_name(company, region):
+    """Test getting the AWS region name."""
+    provider = BraketProvider()
+    fake_arn = f"arn:aws:braket:::device/qpu/{company}/device"
+    assert provider._get_region_name(fake_arn) == region
+
+
+@pytest.mark.skipif(skip_remote_tests, reason=REASON)
+def test_job_wrapper_type():
+    """Test that job wrapper creates object of original job type"""
+    device = device_wrapper("aws_dm_sim")
+    circuit = Circuit().h(0).cnot(0, 1)
+    job_0 = device.run(circuit, shots=10)
+    job_1 = job_wrapper(job_0.id)
+    assert isinstance(job_0, job_1)
+    assert job_0.vendor_job_id == job_1.metadata()["vendorJobId"]
